Rename PhoneNumberVerificationForm -> AuthDetailsVerificationForm
- as we also verify email if the user notification method is set 'email'
